Move events.js to helpers directory to align with dom.js
